requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
crewai
langchain-openai
openai
pandas
numpy
yfinance
textblob
streamlit